Refactor our fetchData function to work w/ redux and to be called when App mounts
