+++ lib/components/DevList.js
@@ -2,22 +2,12 @@ import React, { Component } from 'react'
 import firebase, { reference } from '../firebase';
 import NavBar from './NavBar';
 import { map } from 'lodash';
+import userContainer from '../containers/userContainer';
 
-export default class DevList extends Component {
-  constructor() {
-    super()
-      this.state = {
-        userData: []
-      }
-  }
-
-  componentDidMount() {
-    this.fetchData()
-  }
-
+class DevList extends Component {
 
     renderListings() {
-      return this.state.userData.map(user => <article className="listing">
+      return this.props.userData.map(user => <article className="listing">
                                                <p>
                                                  {user.name}
                                                </p>
@@ -33,16 +23,6 @@ export default class DevList extends Component {
                                          </article>)
     }
 
-  fetchData() {
-    let convertedSnapshot = [];
-    reference.limitToLast(100).on('value', (snapshot) => {
-      let userData = snapshot.val() || {};
-      map(userData, (val) => convertedSnapshot.push(val))
-      let users = convertedSnapshot.map(user => user)
-      this.setState({ userData: users })
-    });
-  };
-
   render() {
     return (
       <section className='dev-list'>
@@ -52,3 +32,5 @@ export default class DevList extends Component {
     );
   }
 }
+
+export default userContainer(DevList)
