+++ lib/components/Map.js
@@ -1,19 +1,8 @@
 import React, { Component } from 'react'
 import axios from 'axios'
-import firebase, { reference } from '../firebase';
 import userContainer from '../containers/userContainer';
-import { map } from 'lodash';
 
 class Map extends Component {
-
-  componentDidMount() {
-    this.fetchData()
-    let map = new google.maps.Map(this.refs.map, {
-      center: {lat: 42, lng: -97},
-      zoom: 4,
-      mapTypeId: 'terrain',
-    });
-  }
 
   componentDidUpdate() {
     let map = new google.maps.Map(this.refs.map, {
@@ -39,16 +28,6 @@ class Map extends Component {
     })
   }
 
-  fetchData() {
-    let convertedSnapshot = [];
-    reference.limitToLast(100).on('value', (snapshot) => {
-      let userData = snapshot.val() || {};
-      map(userData, (val) => convertedSnapshot.push(val))
-      let users = convertedSnapshot.map(user => user)
-      this.props.getAllUserData(users)
-    });
-  };
-
   render() {
     return (
       <section className='map-container'>
@@ -63,4 +42,4 @@ class Map extends Component {
   }
 }
 
-export default userContainer(Map);
+export default userContainer(Map)
